#288 remove redundant service class
